Fixes https://github.com/src-d/rovers/commit/bf7094e6f50937ef94075a5705cda1de51d352f9
Apparently `&quot;` inside comments is replaced by `"` which is not ideal if those comments contain JSON and you need to parse it. This is all done by the official html5 parser of Golang so it affects any libraries depending on it, like goquery. Problematic name examples: `"John \"Nickname\" Doe"` which is encoded in the comment as `"John &quot;Nickname&quot; Doe"` and when we decode we finally get `"John "Nickname" Doe"` which of course makes the JSON parser cry, a lot.