watcher: Exponential backoff when beanstalk fails
When beanstalk fails for different reasons (lost connection, memory is full, etc.) we need a way to handle this kind of errors to recover them eventually.

- Changed beanstalk client library to lentil. The old one do not handle correctly connection problems.
- All the logic is implemented into beanstalkQueue
- Removed unused write mehtods into beanstalkQueue

The old library is already in use into cmd_repo_providers_test file, because we want to test if a job is correclty serialized and obtained for the Fetcher. Fetcher uses the old library.