providers.cgit: save all obtained urls.
In a project cgit page we may have more than one git url.
We will save this urls in a field called 'Aliases' in the cgit collection and also into the repositories.Raw object.

Also I added a test to check if the new logic used to get the main url is the correct one.