cgit: New collection with all discovered cgit urls
Added a new collection to set all the main cgit pages discovered. Now, after call to the Searcher, we check if the result urls are cgit urls and save them.
When rovers is restarted, now we initialize the cgit scrapers using this new collection instead the repositories one.

Also, to be able to check if an url is a cgit one or not, we created two utils files into utils package.

The goquery.go file is able to provide a goquery document using a http.Client with a correct timeout.
The url.go file has some URL transformation methods useful for some transformations between url.URL and string.