Use Mention model from core
Change the old repositories.RAW model to models.Mentions one. Also added all the kallax tags to be able to save them into postgres.
We also replace the github API library for a custom client that does specifically the things that we need. This change is to be able to add kallax tags and have the control of the github repository model.