github: Add replayer
We need a way to add again to the rovers queue all the mentions that we found on the past.
To do this we add a new concept called replayer.
This tool will iterate all info from a specific provider and create again the previous Mentions.

When all the elements on the database are translated to mentions, the replayer shut itself down forever.