github: handle rate limit errors gracefully
When we find a rate limit error (403 code + X-RateLimit headers), instead
of failing, we wait and retry. Usually the rate limit error shouldn't be
hit.

This happens only if more than one process is accessing GitHub API
with the same token (as it happens with CI, but also in other cases).

Signed-off-by: Santiago M. Mola <santi@mola.io>